jira-ai-poc/agent/requirements.txt
# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# HTTP client
httpx>=0.25.0

# Environment
python-dotenv>=1.0.0

# LLM & Embeddings
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-gigachat>=0.2.0

# Vector store
faiss-cpu>=1.7.4

# Embeddings model
sentence-transformers>=2.2.0

# Utilities
jinja2>=3.1.0
pyyaml>=6.0
